移除无用代码；答题影响数值待修复。
